(arcan-net) netpipe test mode added
This pushes multiple cleanups in the networking proxy implementation.
It should be 'working' enough to start moving events back and forth, but
something fails in the REGISTER/ACTIVATE stage on the way, leaving the
client dangling in video-buffer commit.

It is likely that this is somewhere in the arcan<->shmif<->shmif-server
implementation that needs to be stepped through in a different env. as
the same sortof behavior persists in tests/core/proxycon.